refactor: improve CLI UX and code organization
- Split monolithic functions.ts into focused modules (download, env, extract)
- Replace global verbose state with module-scoped setter
- Add interactive project-type selection when --type is omitted
- Validate project names to reject invalid folder characters
- Improve spinner feedback with separate steps for download and project creation
- Replace default exports with named exports (runWebPostSetup, runNativePostSetup)
- Add URL encoding for branch names
- Simplify git initialization with try/catch instead of Future pattern
- Enhance README with direct links to web/native repos and collapsible troubleshooting
- Add lint & format check step to build workflow
