+++ .github/workflows/build.yml
@@ -25,6 +25,9 @@ jobs:
 
       - name: Install dependencies
         run: pnpm install
-      
+
+      - name: Lint & format check
+        run: pnpm check
+
       - name: Build cli
         run: pnpm build
