+++ src/lib/debug.ts
@@ -1,11 +1,13 @@
 import chalk from 'chalk';
 
-/**
- * function used to log debug message only if --verbose flag is used
- * Act as a wrapper around console.log
- */
+let verbose = false;
+
+export const setVerbose = (value: boolean) => {
+  verbose = value;
+};
+
 export const debug = (...args: unknown[]) => {
-  if (global.isVerbose) {
+  if (verbose) {
     console.debug(
       `\n${chalk.magenta('DEBUG')}\n`,
       ...args.flatMap((valueToDisplay) => [chalk.magenta('|>'), valueToDisplay, '\n']),
